:truck: chore(api) move `http-agent` to `src/shared/'
