+++ api/src/prescription/campaign-participation/domain/usecases/send-shared-participation-results-to-pole-emploi.js
@@ -1,5 +1,5 @@
 import * as httpErrorsHelper from '../../../../../lib/infrastructure/http/errors-helper.js';
-import { httpAgent } from '../../../../../lib/infrastructure/http/http-agent.js';
+import { httpAgent } from '../../../../shared/infrastructure/http-agent.js';
 import { logger } from '../../../../shared/infrastructure/utils/logger.js';
 import { PoleEmploiPayload } from '../../infrastructure/externals/pole-emploi/PoleEmploiPayload.js';
 import { PoleEmploiSending } from '../models/PoleEmploiSending.js';
